+++ src/components/TextInput/TextInput.tsx
@@ -1,106 +1,101 @@
 import * as React from "react";
-import bem from "../../utils/bem";
+import {
+  Box,
+  Input as ChakraInput,
+  Textarea as ChakraTextarea,
+  useMultiStyleConfig,
+} from "@chakra-ui/react";
+
 import { TextInputTypes, TextInputFormats } from "./TextInputTypes";
 import Label from "../Label/Label";
 import HelperErrorText from "../HelperErrorText/HelperErrorText";
 import generateUUID from "../../helpers/generateUUID";
 
 export interface InputProps {
-  /** Additional attributes to pass to the `<input>` tag */
+  /** Additional attributes to pass to the `<input>` or `<textarea>` element */
   attributes?: { [key: string]: any };
-  /** HTML Input types as defined by MDN: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input */
-  type?: TextInputTypes;
-  /** Will add 'aria-required: true' to input */
-  required?: boolean;
+  /** Populates the HelperErrorText for the standard state */
+  helperText?: string;
+  /** ID that other components can cross reference for accessibility purposes */
+  id?: string;
+  /** Populates the HelperErrorText for the error state */
+  invalidText?: string;
+  /** Adds the `disabled` and `aria-disabled` prop to the input when true */
+  isDisabled?: boolean;
+  /** Adds errored styling to the input/textarea and helper text elements */
+  isInvalid?: boolean;
+  /** Will add `required` and `aria-required` props to the input/textarea elements */
+  isRequired?: boolean;
   /** Provides text for a `Label` component if `showLabel` is set to true;
-   * populates a `aria-label` sttribute if `showLabel` is set to false. */
+   * populates an `aria-label` attribute if `showLabel` is set to false. */
   labelText: string;
+  /** The action to perform on the `input`/`textarea`'s onChange function  */
+  onChange?: (
+    event:
+      | React.ChangeEvent<HTMLInputElement>
+      | React.ChangeEvent<HTMLTextAreaElement>
+  ) => void;
+  /** Populates the placeholder for the input/textarea elements */
+  placeholder?: string;
   /** Offers the ability to show the label onscreen or hide it. Refer to the
    * `labelText` property for more information. */
   showLabel?: boolean;
   /** Offers the ability to show the "Required"/"Optional" label onscreen or
    * hide it. True by default. */
   showOptReqLabel?: boolean;
-  /** Populates the HelperErrorText for standard state */
-  helperText?: string;
-  /** Populates the HelperErrorText for error state */
-  invalidText?: string;
-  /** Populates the placeholder of the input */
-  placeholder?: string;
-  /** Populates the value of the input */
+  /** HTML Input types as defined by MDN: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input */
+  type?: TextInputTypes;
+  /** Populates the value of the input/textarea elements */
   value?: string;
-  /** className you can add in addition to 'textinput' */
-  className?: string;
-  /** ID that other components can cross reference for accessibility purposes */
-  id?: string;
-  /** Helper for modifiers array; adds 'errored' styling */
-  errored?: boolean;
-  /** Adds the 'disabled' prop to the input when true */
-  disabled?: boolean;
-  /** Modifiers array for use with BEM. See how to work with modifiers and BEM
-   * here: http://getbem.com/introduction/ */
-  modifiers?: string[];
-  /** The action to perform on the `<input>`'s onChange function  */
-  onChange?: (
-    event:
-      | React.ChangeEvent<HTMLInputElement>
-      | React.ChangeEvent<HTMLTextAreaElement>
-  ) => void;
 }
 
-/** The type used for `ref`s. We want to extend both `input` and `textarea`
+/**
+ * The type used for `ref`s. We want to extend both `input` and `textarea`
  * since both are available to create through `TextInput`.
  */
 export type TextInputRefType = HTMLInputElement & HTMLTextAreaElement;
 
-/** Renders either an `input` element with a specified type or a `textarea`
- * element. All types will render an accessible Label component and an optional
- * HelperErrorText component.
+/**
+ * Renders either an `input` element with a specified type or a `textarea`
+ * element. All types will render an accessible `Label` component and an
+ * optional `HelperErrorText` component.
  */
 const TextInput = React.forwardRef<TextInputRefType, InputProps>(
   (props, ref: React.Ref<TextInputRefType>) => {
     const {
       attributes = {},
-      type = TextInputTypes.text,
-      required,
+      helperText,
+      id = generateUUID(),
+      invalidText,
+      isDisabled = false,
+      isInvalid = false,
+      isRequired = false,
       labelText,
+      onChange,
+      placeholder,
       showLabel = true,
       showOptReqLabel = true,
-      helperText,
-      invalidText,
-      placeholder,
+      type = TextInputTypes.text,
       value,
-      className,
-      disabled,
-      errored,
-      onChange,
-      modifiers = [],
-      id = generateUUID(),
     } = props;
-
+    const styles = useMultiStyleConfig("TextInput", {});
     const isTextArea = type === TextInputTypes.textarea;
     const isHidden = type === TextInputTypes.hidden;
-    const optReqFlag = required ? "Required" : "Optional";
-    const errorOutput = invalidText
+    const optReqFlag = isRequired ? "Required" : "Optional";
+    const finalInvalidText = invalidText
       ? invalidText
       : "There is an error related to this field.";
-    let footnote;
+    let footnote: string | React.ReactNode = isInvalid
+      ? finalInvalidText
+      : helperText;
+    let fieldOutput;
     let options;
 
-    if (!showLabel) attributes["aria-label"] = labelText;
-    if (helperText) attributes["aria-describedby"] = helperText;
-
-    if (errored) {
-      modifiers.push("error");
-    }
-    if (isTextArea) {
-      modifiers.push("textarea");
-    }
-
-    if (errored) {
-      footnote = errorOutput;
-    } else {
-      footnote = helperText;
+    if (!showLabel) {
+      attributes["aria-label"] =
+        labelText && footnote ? `${labelText} - ${footnote}` : labelText;
+    } else if (helperText) {
+      attributes["aria-describedby"] = `${id}-helperText`;
     }
 
     if (
@@ -120,11 +115,11 @@ const TextInput = React.forwardRef<TextInputRefType, InputProps>(
 
     options = {
       id,
-      className: bem("inputfield", modifiers, "", [className]),
-      "aria-required": required,
+      "aria-required": isRequired,
       "aria-hidden": isHidden,
-      disabled,
-      required,
+      isDisabled,
+      isRequired,
+      isInvalid,
       placeholder,
       onChange,
       ref,
@@ -134,38 +129,37 @@ const TextInput = React.forwardRef<TextInputRefType, InputProps>(
     // also needs `type` and `value` to render correctly.
     if (!isTextArea) {
       options = { type, value, ...options } as any;
+      fieldOutput = <ChakraInput {...options} __css={styles.input} />;
+    } else {
+      fieldOutput = (
+        <ChakraTextarea {...options} __css={styles.textarea}>
+          {value}
+        </ChakraTextarea>
+      );
     }
 
-    const fieldOutput = React.createElement(
-      isTextArea ? "textarea" : "input",
-      options,
-      isTextArea ? value : null
-    );
-
-    const transformedInput = (
-      <div className="textinput">
+    return (
+      <Box __css={styles}>
         {labelText && showLabel && !isHidden && (
           <Label
             htmlFor={id}
             optReqFlag={showOptReqLabel && optReqFlag}
-            id={id + `-label`}
+            id={`${id}-label`}
           >
             {labelText}
           </Label>
         )}
         {fieldOutput}
-        {((helperText && !isHidden) || errored) && (
-          <HelperErrorText isError={errored} id={id + `-helperText`}>
-            {footnote}
-          </HelperErrorText>
+        {footnote && (
+          <Box __css={styles.helper} aria-disabled={isDisabled}>
+            <HelperErrorText isError={isInvalid} id={`${id}-helperText`}>
+              {footnote}
+            </HelperErrorText>
+          </Box>
         )}
-      </div>
+      </Box>
     );
-
-    return transformedInput;
   }
 );
-
-TextInput.displayName = "TextInput";
 
 export default TextInput;