+++ src/theme/index.ts
@@ -11,6 +11,7 @@ import Radio from "./components/radio";
 import CustomRadioGroup from "./components/customRadioGroup";
 import CustomCheckboxGroup from "./components/customCheckboxGroup";
 import Tabs from "./components/tabs";
+import TextInput from "./components/textInput";
 import CustomSelect from "./components/select";
 import Heading from "./components/heading";
 
@@ -52,6 +53,7 @@ const theme = extendTheme({
     CustomRadioGroup,
     CustomSelect,
     Tabs,
+    TextInput,
   },
   // Chakra prefixes its own CSS variables with `--chakra` by default but this
   // can be updated to be anything we want. This can be "nypl" to have the